Added Beta3 and Beta4
Beta3 was where I started a rewrite of the converter to fix a nasty race condition bug Chase discovered but it's going to end up in a complete rewrite of most of the logic. As such, Beta4 was forked as the real "release" and Beta3 will eventually turn into Convertify 1.4 in the future.